Use quadratic concatenation for C conway.
Use the same "quadratic string concatenation" algorithm in Lua and C
implementations of this benchmark.
